Move Idena bot banner to Profile
Made idena bot banner self-containing
